backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
alembic==1.13.1
pydantic==2.7.1
pydantic-settings==2.2.1
python-dotenv==1.0.1
celery==5.4.0
redis==5.0.4
pyjwt==2.8.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.9
cryptography==42.0.5
httpx<0.28.0

# RAG & Documents
chromadb==0.5.0
groq==0.5.0
sentence-transformers==2.7.0
PyMuPDF==1.24.4
python-docx==1.1.0
python-pptx==0.6.23
langchain==0.1.20
